LayoutTests/fast/viewport/viewport-warnings-3.html

<head>
    <title>Unrecognized value.</title>
    <meta name="viewport" content="width=unrecognized-width">
    <script>
        function test() {
            if (window.testRunner) {
                testRunner.dumpAsText();
                alert(internals.configurationForViewport(document, 1, 320, 480, 320, 352));
            }
        }
    </script>
</head>
<body onload="test();">
